my_ai_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
ultralytics==8.0.201
torch==2.1.1
torchvision==0.16.1
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
python-multipart==0.0.6
